Testing storing chunk data in resources
To make setting up maps easier, added Level_Data to hold what types of chunks to load and Chunk_Data to store additional spawn info for the chunks.
The goal now is to have one handle for world and instances, but the level_data will deiced on if it load like an endless world or limited to static chunks.