update
added an event object(resource) to be used later as a way to share data for all listening to an event(or signal).

testing with noise. may store world gen noise in level data or change it to a world data object and expose it to the logic that dose the generation.