update
Started to change the Generic_Event to be a static class and change it to Event, but may remove it or not use it as much. The orignal goal was the abilty to cancel events or signal by those who were connected. This is not really ideal to focus at this time.

added more signals to try to connect things better. Player handler, Inventory, and Fish gain a few to try to help with updating the score and later the HUD.