update
Made it so the handler can build themselves up as well as added Game_Handler as an autoload. The Main scene may contain a start menu or some sort of interface to communicate with the game if it have not built all of the systems, but for now the game starts up fully.