minor update
change how item data is created by not using a static creation function. It type is now the resource used to generate it.

started working on an action system to replace or improve upon the interaction system. It will be similar to the item system in the way it will use static function and dictionary. A action handler will be use to run and clear the action as needed.