Started moving logic to Interaction_Data resource.
Trying to add a plug and play with actions tie in with global static functions for out of scope logic which will be located in the event folder.

Added a features folder for systems that meant to be independent of the game. This is where the interactive data is stored.

Also removed the .wave music so file size stay small.