Minor update
I change how the movement system works. It will use a resource base system called components (like the interaction system). This will allow different movement logic that can be pick from the editor instead of characters having to implement similar, yet different, logic.
CharacterMovement will be rename when an idea for what will suit it best for the base of a character that can move (or is controlled by player or ai)