minor change
Added a Game_Ulitilty script for static helper functions.
The current use is to handle the on_interact function abstractly in the player/ai handler. Mostly because I may experment of using metadata to pass actions like on_interact and it may be easier to have that logic in one place and I am not sure if I want AI and Player handler to share a child controller handler yet. Other use is to have logic for game stored in one place without ties to a sole handler.