update
working on Region_Data Resource which is meant lightly store tile data for a tilemap. It meant to make loading dynamic tilemaps easier by supporting chunking the region. The idea is to run the logic in regional data and then render it by chunk as needed. Also allow the map to be save, load, or pass around since it just a nested array. The logic should be push to a lower level, but the gdscript should work just fine.
Note: this data is to reduce the need for tilemaps nodes to be loaded to make use of. If necessary, the data could be use to hold template for tilemaps without being dependent on a single tilemap. The currently tilemap seem to lack some feature exposed to gdscript and that the only reason for the Regional data.
The work on remaking the tilemap world have now begin (yet again) in hope to have bigger tilemaps with faster loading. Maybe with tilemap recyling to reduce creation overhead.