update
Rearrange some things with multiplayer in mine despite not part of the scope. Also added an input_handler so the player_handler did not need _input. It also was made with the idea of limiting input to one player handler since I an not sure how to disable built in listeners per class. So in theory, the game can turn off input to the world and additional reroutes could be made.
alsp decided to use autoloads so I do not need to store ref in objects of the world_handler or hud. Most now grabbing from the old Game autoload via static function I should move to the game_handler.