minor progress
Prepping to rework the world handler to make working on it easier.
Also split the water in a different tileset so it is easier to work with, but lack variation to use as a proper terrain.
Also move some event functions into the game_handler so that they are all in one place and can be called via Game autoload.