update
made a dedicated generator_data so that generator logic can be added to objects. It can generate on a Node or an Array. The later is for cases where more than one generator may need to run and need a simple type to acess. It also have related signals to notify owners when they are finish in the case the generator takes more than one tick.