update
decided to bring back the use of static libraries for some major events/calls since there seem to be a delay in the signals. The idea is attachable actions, debugging, and maybe handlers will call this to trigger an event sequence.

Also manage to add back fix position static tilemaps to the world gen as well save the world foliage generator so these tilemaps can use them to match with the world.