minor cleanup
Added static function libraries to handle Metadata and Savable_Data abstraction to help with debugging and refactoring.
Also have HUD listen to player state changes. Currently need to redesign how the Score is updated since now the order of the score is not fixed.
Rename and move the starting map in it own folder and planning to have a world folder and the files name as [region_name]_x[x]_y[y] or similar.