update
remove the Timers_Handler since having each handler owning a timer(if needed) may be better and easier to get connections right.
Work on the save system and added a Resource_Handler for catching some data and storing paths. It is currently used to store client settings (currently music volume) and handle saving and loading it since it would be redundant to use ConfigFile with SavableState. The idea is to allow other gui elements(or any nodes) get access to the configs in one spot with little navigation through object. Mostly I did not want to have to keep opening and saving the file.