update
Move more handler to be autoloads.
Also made a line trace resource for Raycast that do not need to happen often. It also have multi object raycast and I am debating if I should add shape and point cast to it as well or if those cases best serve with the Raycast node.
Also the player state currently saving because I forgot to disable it. It need a uid. Also game saves should be stored in folders so more than one state can be save. Still need to make it so new game will clear the state. May need player handlers in a group (or loop the array) and tell them to load (and when to save) maybe even to reset themselves.