update
Merge Hud and Input handler into UI_handler singleton(UI) since this node will handle the flow of the user input. Hud would more or less be a child of this node.
Also decided to use ConfigFile for save data since it can easily be made hard to read/edit and it follow a similar format to the savable state data. Still deciding if I want to keep that dictionary to support other save types or depend fully on ConfigFile.
Tested the inventory with the save system. currently using the item type as a UID instead of an object ref. Need to see if there an overhead with loading the type and if so see if ResourceLoader have a feature to get loaded resource from an id else add that feature to Resource_Handler