update
prototype a new dialog system, but may not use it in favor of an action system.
Also prototype an action system that support conditions. It should help with interacting with system without the need to make many scripts. It not a replacement to scripting, but a way to reuse common actions and checks within the editor.