minor update
Decided to have the controller call raycast on the pawn instead of handle it self. It will listen for the response and execute the action if there is any allowing the pawn more control.
I am also thinking about eventually to move from the interactive data and just use node like componets that call actions to start up or modify gameplay systems. An example is the fishing game will work like the dialog system. Currently the interactive data running most of the logic, but can be simplified to run an action on cancel, on catch, and on failed catch if needed.