update
added controller support for the menus.

I mess with saving resources in the savable states and it seems to work(with the config save type) with the exception of needing the force the state to be flag dirty. I may change the items to be a resource instead of a dictionary in the future.