update
Mess around with the Character2d trying to see what can be made reusable and understanding how to design characters easier.
Decided to use Shapecast for interaction and attack checks if provided(just need to move it from player to character2D) and area2d could be used with local logic if null.

Also mess with animator player to make the cat bounce as it walks and sit when idle.