minor update
Messing around with having a brain for the character and how it should be handled. Currently the player brain get a move direction from input and the bat brain move to a point set by the controller. I also want to have a object the brain can share that can be used to decide stuff it may not have access to, but the data available is base on what the AI controller wants to add. I may use metadata for this since it could be too abstract to design without typed dictionaries.