minor update
I decided to convert level_data(Resource) to Base_Level(of type of at least a Node). The reason is that it should allow making levels easier without making a tool. Also I realize that it may be easier for them to handler their children in the future and that the worlds handler just an interface to the current active level. In the future this would help keeping levels in memory to allow swapping back to it quicker. Currently forgot to add the way to queue free unused levels, but may use the array and push out the oldest(if not the current one loaded) if the array is too big and change this logic if the base_level ever turn into a viewport(isolated world2d).