minor update
I made an inventory handler that will handle the bulk of the item system. I want it to be a module for other project so I felt this will help make it easier to connect to and communicate to. Also can be an autoload like node or use as a static library(but the add item logic would need to be added else where with the static approch)