update
Some old stuff I forgot to push. Currently slowly changing they system to be more simple. Added a Data_Handler to replace Resource_Handler. It will handle loading and saving as well as provide links to data in the res or user path if needed. Also clean up the item system a bit.