update
Change more of the flow. Characters will have a state that can be saved to the save data. This also means handlers should not be passed in actions and such since the pawn should have the data needed.

Detach the Player Handler from the Game Handler so that the Player can be accessed globally. This means the Player will be the client controlled player. If networking is added, the game would update the state and characters directly.

The new flow is to try to have nothing access the Game Handler and let it act as a manager of the other handlers and maintain the flow of the game. The other handler should only care about themselves and their children, but anything can access them.

Also levels will be added as a current scene of the scene tree. This means anything scene can be loaded, though they may need to the other handler so they do not behave improperly.

I need to move pawn input back to the player handler instead of the UI still as well as a lot of other clean up.