update
I started working on making the game handler manage more of the persistent state and other handlers.

The UI was worked on a bit, but it still need to stop accessing other handler. I could cheat and move the debug scrip into the debug child, but that for another day

Remove some old scripts that are no longer used. Also remove the state in the game handler and used the data save state instead for the save game seed. (I should remove the player handler state since the data handler should store savable player data as well).

Also worked on the character and brain. Player handler will set the brain direction and the character will get the direction(by passing velocity and position) every time move() is called. The brain can decide on deviation of the direction from the given data and any other data that may be added to it (such as the AI brain move_to location (which currently added by metadata))