update
started working on the switch to tilemap layers. The generator is using it, but it should try to handler it as a multi layer instead of single. I did add a concept node that would act like it, but have not made full use of it.

Also did some clean up and added notes to the todo_and_info text about the flow and structure. There still a lot more clean up to be done.