minor update
Started changing how the pawn is handle. Will use a brain (kind of like a controller) that the AI or Player handler can use to take more control of or give it data to use. I also decided to not store a pawn ref in the Player Handler, and instead store a brain that can be assign to pawns for direct input.

This also means a lot of test broke since pawn in Player Handler is no longer a thing. Groups are being used to get ref to player pawn and enemies, though some use cases are temp to make sure it is work.

I have not fully check if everything working as it should.

Also I change fetching of data for the score...but it seems it not working as expected. I try to make it listen to the savable state and update to changes beside updating on saving, but it only updating on saving. I just need to look into it again. I just do not want to use processes to update UI as long as I do not have to.