update
I decided to use the brain/controller as a shared object between the hander and pawns. This allow the resource to be assign in the engine and connect the handler to the pawn as long as both are using the same instance. This means the get_direction would need a data parameter for more complex cases where the pawn need to make decisions base personal interests. I plan in using an object or resource instead of a dictionary in the future.

Also did some clean up.