minor update
change the action system to use an object instead of a dictionary for the action state. Also tried to detach the need for a handler.

Nodes that use Actions should try to pass the object(Action_Data) with a valid owner (which should be the node that owns it). The target is if the effect of the action is meant to happen to a different known node and should also be pass if the event has one. Extra data can be stored in the metadata of the passed state or the state can be extended.

Not much and more cleanup to do when I find the time to do so.