minor update
Clean up the generation stuff a bit. Also using a shared resource for the world environment state so the world handler can forward time and the levels can listen for the change. I should remove environment color so the level can use it own...or I can keep and let the level use their own instead if necessary. That resource still need to be design better, but some kind of "time" resource will be used for time stuff.