minor update
I look into the generator system and decided the resource was doing too much. It should only provide method for getting data at a point as well as hold the data sources.

I also made a Tile_Info to store info about tile such as tile id and atlas coord. This mostly so there only one data per tile type that can be used when ever one need to check about a tile and I am not sure of any built in way that do.