minor update
Note: saves will break, though saves were not expected to be safe in dev.
(todo: decided to fine a safe way to ref resources or be diligent with file dir)

Major change is reorganizing the file structure. The plan is to split it into three major dir for engine tools, common base scripts and resources, and game data.
bulk of the game data will be in the Game/Data. Common may need to be redesign, but Common/data currently where the bulk of the data will be. I may pull features from it and put it under Common and reserve data for base data not group together, but not sure yet.