update
started changing the save system a bit.
The savable state will handle more if the save and load logic but the main handler of the state need to call them.

I wanted to not save the whole state, but only the meta and values deems worth saving. Most of all I wanted to update the state instead of replacing it with a new one. Mostly because I want the tres resources in the filesystem to load the save data the player saved. This is to have states being able to be drag into the slots. I do need to add a a way to flag it to be dup for states that are meant to be static (like enemies) also a way to set the save_id to the enemy id...though such states should not have a need to be saved.