minor update
modified the save state a bit. I set up a system that use the scene tree path as the save id if the state enable it. It should default back to the id(if set) if the path is not set. This is more for states that represent the state of a scene/node. This only will work for character added to a scene. Dynamically added character should have a handler that keep track of what it creates and load/save their state to itself.

Also added a get_level_id to world so there a safer way to get some kind identifier. I still need to add exit info to the world.