minor update
This mostly a backup before I change how character states are handled. Any change at runtime will be stored in a object(savable state) to be used for saving and loading. Another resource will be used for base stats. The solution is simple, but the state is still in a mess and will break a lot of stuff untill I clean it up.