Update
Started on the transfer. Character now can handle loading and saving it own state, but may need to override the get_save_path depending on where it need to save. It using a path base on where it exist in the scene tree, but player character override it to be in a fix subdirectory called characters. I am also trying to keep base classes from depending on autoloads or global variables that is not part of the base engine.