Update
I started moving to an resource based event system to work parallel with the states. The plan is to move away from the global state and groups for the base game and features and use shared objects and resources instead for communication. Groups will be reserved for game content or calling to  nodes that do not share a parent class.