update
Added a rough way to load music. This is because I want to not include certain assets in github if I do not own it and not sure if the copywrite support it.

Also it will load from users side so an early form of music modding. I will move and clean it up once I know where and how the sound will be handle.
For now any vaild music resource under Game/Data/Assets/Music that start with background will be used as background music. for users, it will pull from a Music dir(caps matter for now) prefix with 'background' and is a mp3 file (have not supported the others  yet).