minor update
Moved the files getter to the data handler. Also convert it to a static function so it could be reused (though I will try not too for most nodes). I may move all runtime resource modification here or to a child nodes if it get too big. Mostly states, music, and probably art assets will be patched. I may allow static states to be patched, but for now I rather not have gameplay be affected for this project.