minor update
Moved the player handler to be a child of the game handler instead of an autoload. Will try to do that to the other autoloads if they need to be a scene, else I'll try to convert them to a resource managed by another hander.

Also ran into saving issue, but I think I solve it. I do need to make a resource to share path data since the character needs it to load its save data. (so data handler may become a resource since the need to be a scene is to call group stuff for saving, but as a resource it would provide them as a signal with type check features)