minor update
created a data field in the player state since the savable_state may be changed to a state since saving is handle by the handler using some resource or object to put and pull save data from. So the savable state may be used, but rename to be less misleading (it will only have a easy way to add and get data in the form of an dictionary). I would use metadata, but it seems like it would not save or load correctly on old test and I do not want to test it again.

also made the music louder at max since it seem softer than other sounds.