minor update
slowly trying to clean up saving and loading. The plan is to use a dictionary to store all savable data and then store it  as or in a serialable object such as a resource, config file, or something.

The state will handle getting and setting from this dict while some handler will handle saving and loading the dict. This is mostly because most states have a fix path I have trouble reassigning without breaking something so I find it easier to bundle the data in a simple abstract (dict but array could work if order never changes) object to be handle. Also should make loading as a non-resource type easier to add as long as custom godot types are handled (vectors and converting resource ref to uid or paths)