minor update
clean up the world handler a bit and decided to use the save state the game state is using to store both the game and world state. I may do that for most states handle by the game.

Also decided to store the level uid in the world state so level that the last save happen in will be loaded. Just need to check to see if the default is set if modified in the editor when I feel like it.