minor update
More clean up while trying to decided how to handle the interaction system. Will keep the lastest way, but will depend on the component to provide additional data (later) like target character state. (also I need to remember to decide if I should pass the collision results as additional data). also the additional data is a bit odd. It move about as a dictionary, but get set as metadata in the action state. This could work if the state do not get saved.