minor update
Working on moving the inventory to the player state and having two types. One like old jrpg inventories and another for limited slot for items with metadata (current system use this approch). The plan is to store the data in the state as dict, arrays, and basic types to help with future saving cases. There is little reason to store it as an object when it is not actively being used or modified and converting it back and forth should not be that much work.

Also this means item_type will have all the static data about the item so that item will act as the item's state. Old approach had some static info in the item resource.