minor update
Connected and tested the new place for the advance items, but still have some other not really used parts to update. It is a bit odd converting between an object and a dictionary, but it should help convert to json if I decided to use that as a save format. (I like saving as a resource, though a bit redundant with the current approach. Too many people worry about security so I have to keep other options in mind).