minor update
I forgot to update the pass item amount in the advance inventory case. It is currently unused, but the plan that the item being added amount will change near 0 base on how much is added or remove. Anything not Zero means the item was not used up or the player did not have enough of that item to give up. It felt redundant to pass and return an amount when an object with same values is being pass. If the item needs to be read only for some reason, then it can be duplicated instead and use that version to check for the difference in state if needed.